Top 25 Wikipedia
Edited script to only pull the top 25 dates a Wikipedia page was viewed. The script will prompt the user to input what they'd like to search for and how many years. Afterward, it will create a CSV containing the dates visits of the page's top 25 days within the time period the user selected.